requirements.txt
gradio==5.22.0
icecream==2.1.4
torchvision==0.21.0
controlnet_aux==0.0.9
mediapipe==0.10.21
diffusers==0.32.2